+++ client/App.js
@@ -20,7 +20,7 @@ import Welcome from "./components/Welcome";
 import EditUserInfo from "./components/EditUserInfo";
 import ChangePassword from "./components/ChangePassword";
 import ADdashboard from "./components/ADdashboard";
-import AddProduct from "./components/Addproduct";
+import AddProduct from "./components/AddProduct";
 
 import { me } from "./store";
 import { fetchProducts } from "./store/flowers";
