+++ client/components/ProductInfo.js
@@ -15,6 +15,7 @@ import CheckBoxOutlineBlankIcon from '@mui/icons-material/CheckBoxOutlineBlank';
 import { Box } from "@mui/system";
 import Fab from '@mui/material/Fab';
 import AddIcon from '@mui/icons-material/Add';
+import "./admin/productInfo.css";
 
 class ProductInfo extends React.Component {
   constructor(){
@@ -110,7 +111,7 @@ class ProductInfo extends React.Component {
     const icon = <CheckBoxOutlineBlankIcon fontSize="medium" />;
 
     return (
-      <div>
+      <div className="productList">
         <Box sx={
           {
             display: "flex",
