+++ client/components/admin/sidebar.css
@@ -1,7 +1,7 @@
 .sidebar {
     flex: 1;
-    height: calc(100vh - 50px);
-    background-color: rgb(251, 251, 255);
+    height: calc(69vh - 85px);
+    background-color: #dab3c0;
     position: sticky;
     top: 50px;
   }
@@ -17,7 +17,7 @@
   
   .sidebarTitle {
     font-size: 13px;
-    color: rgb(187, 186, 186);
+    color: #fff;
   }
   
   .sidebarList {
@@ -31,7 +31,7 @@
     display: flex;
     align-items: center;
     border-radius: 10px;
-    color:rgb(44, 44, 44) 
+    color:rgba(0, 0, 0, 0.9) 
   }
   
   .sidebarListItem.active,
